Setting up initial repo
We are taking in the code that Claudia emailed to us on 4/13/15 and
creating a new master version that we will be working on. Any old code
is in the ‘OLD’ folder, including the changes I implemented int he
javascript file which I still need to stick into this newer version.
We’re getting there :)