Sticking previous work back in
Taking the boilerplate code written back in March for putting the CSS
stuff into the HTML printout from and sticking it into the current
version.

Also, it seems this code is missing from the April backup for some
reason; I decided to add a warning to the name just in case so I don’t
get confused later on.