Fixing up barcode image
Changed image and the size of the image to make it more realistic.
Should still probably double check with someone to make sure it’ll be
ok.

Also, still need to double check that certain configurations of
selections don’t cause it to run off the page for some reason!