Creating Refrence Backup
Going to start rearragnging how the github branches are set up, I just
wanted to create a backup to reference later. Currently my part of the
code is in a pretty stable place, and other than a problem with the
choices selection (which I assume will be fixed once I merge my stuff
with Matt’s latest changes) everything seems to be working well.