Manual, temporary revert of Matt's commits
Realized that the bug in showing selections may be harder to squash
than I initially thought. As such, for now, I manually ‘reverted’ all
of the code in lines 1-1116 of star.js (‘reverted’ because what I did
was copy and paste in these lines of code) which are the lines that I
was NOT working on for the HTML CSS printout. Doing so fixes the bug in
showing the printout; once we get the bug fixed in Matt’s code we will
accept his changes again.