First commit to ensure syncing
This is a fresh new branch off of Matt's now that we're 100% positive
it's not missing any files. Visual Studio is behind the .cache files
changing, and it's unimportant (no actual changes to the real code in
this commit) but it gives me a chance to ensure the branch is created
and write down some comments.

Possible issues currently:  while it creates the app package correctly
if I right click the StarVote.windows (windows 8.1...) ... and do the
store->createapppackage, it does not work on my machine if I try to do
the same using the project->createapppackage. Something about the images
being messed up, see the package.appxmanifest. Also, something like:

http://localhost:64239 : error  : Error opening web
http://localhost:64239. Unable to open the Web site
'http://localhost:64239'.  The Web server 'http://localhost:64239' could
not be found.

is spat out to the solution output up when I open the solution file. But
regardless things were working apparently? Should look into this later.